refactor: break crypto in separate files
