+++ crates/bls-gadgets/src/hash_to_group.rs
@@ -22,7 +22,7 @@ use algebra::{
     edwards_sw6::EdwardsProjective,
 };
 use bls_crypto::{
-    bls::keys::SIG_DOMAIN,
+    bls::SIG_DOMAIN,
     curve::hash::try_and_increment::get_point_from_x_g1,
     hash::{
         composite::{CompositeHasher, CRH},
@@ -340,7 +340,7 @@ mod test {
     use algebra::bls12_377;
     use r1cs_std::{groups::GroupGadget, test_constraint_system::TestConstraintSystem};
 
-    use bls_crypto::{bls::keys::SIG_DOMAIN, curve::hash::try_and_increment::TryAndIncrement};
+    use bls_crypto::curve::hash::try_and_increment::TryAndIncrement;
     use r1cs_std::bits::uint8::UInt8;
     use rand::{thread_rng, RngCore};
